% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpdata0}
\alias{cpdata0}
\title{A template for carcass persistence data with interval-censored carcass
 persistence times}
\format{
A matrix with 2 columns bracketing persistence times (in days since
 carcass placement) of each carcass:
\describe{
  \item{CPmin}{the last time the carcass was observed}
  \item{CPmax}{the first time the carcass was noted missing}
}
}
\usage{
cpdata0
}
\description{
A template for carcass persistence data with interval-censored carcass
 persistence times
}
\keyword{datasets}
