% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{days0}
\alias{days0}
\title{A template for search schedule data}
\format{
A numeric vector with the times when searches were conducted, with
 days[1] = 0:
\describe{
  \item{days0}{numeric vector of search times}
}
}
\usage{
days0
}
\description{
A template for search schedule data
}
\keyword{datasets}
