% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_CP.R
\name{est_cp0}
\alias{est_cp0}
\title{Fit cp carcass persistence models}
\usage{
est_cp0(cpdata, dist)
}
\arguments{
\item{cpdata}{matrix of interval censored carcass persistence times}

\item{dist}{Name of the persistence distribution family: Weibull,
lognormal, loglogistic, or exponential}
}
\value{
answer
}
\description{
Carcass persistence is modeled as survival function
}
\details{
uses \code{survival} package to fit.
}
