% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_CP.R
\name{getab}
\alias{getab}
\title{retrieve EoA parameterization from \code{survival} parameterization of
 a fitted cp model (or \code{survreg} object with exponential, weibull,
 lognormal, or loglogistic distribution)}
\usage{
getab(cpmodel)
}
\arguments{
\item{cpmodel}{fitted cp model (from \code{est_cp0} or \code{survreg} object)}
}
\value{
2-vector of pda and pdb
}
\description{
retrieve EoA parameterization from \code{survival} parameterization of
 a fitted cp model (or \code{survreg} object with exponential, weibull,
 lognormal, or loglogistic distribution)
}
