% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkdata0}
\alias{pkdata0}
\title{A template for summarized searcher efficiency data with the number of
 carcasses available and the number discovered for N = 12 search occasions}
\format{
A list with 2 numeric N-vectors with numbers of:
\describe{
  \item{n}{searcher efficiency trial carcasses available}
  \item{y}{carcasses discovered}
}
}
\usage{
pkdata0
}
\description{
A template for summarized searcher efficiency data with the number of
 carcasses available and the number discovered for N = 12 search occasions
}
\keyword{datasets}
