% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_CP.R
\name{sim_cp0}
\alias{sim_cp0}
\title{generate random cp parameters or persistence times}
\usage{
sim_cp0(cpmodel, nsim, option = "parms")
}
\arguments{
\item{cpmodel}{Fitted cp model ((\code{survreg} object))}

\item{nsim}{Number of simulation draws}

\item{option}{\code{option = "parms"} returns random draws of parameters from
the fitted model; \code{option != "parms"} returns random draws of carcass
persistence times}
}
\value{
answer
}
\description{
Given a fitted cp model (\code{survreg} object), generate random
 \code{pda} and \code{pdb} parameters or random persistence times.
 NOTE: This function is likely to move to call GenEst's \code{rcp} in future.
 This will not change the results, but the GenEst version is more nicely
 coded and keeping some coherence among the models is helpful.
}
