% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_SE.R
\name{sim_pk0}
\alias{sim_pk0}
\title{Simulate pk parameters from model}
\usage{
sim_pk0(pkmodel, nsim = 1000)
}
\arguments{
\item{pkmodel}{A model returned from \code{est_pk0}}

\item{nsim}{Number of simulation reps for estimating the joint posterior
distribution of \code{p} and \code{k}.}
}
\value{
An nsim x 2 matrix of simulated \code{p} and \code{k} values
 the joint posterior for SE.
}
\description{
Simple simulation
}
