% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_SE.R
\name{summary.estpk}
\alias{summary.estpk}
\title{Summary statistics for estimated p and p parameters}
\usage{
\method{summary}{estpk}(object, crlev = 0.95, n.iter = 10000, ...)
}
\arguments{
\item{object}{An \code{\link[=est_pk0]{estpk}} object}

\item{crlev}{Credibility level of estimated CI to be returned}

\item{n.iter}{Number of iterations of the JAGS model for estimating the joint
posterior distribution of \code{p} and \code{k} (relevant only if
\code{object$type == "pk"}).}

\item{...}{additional (optional) arguments passed to \code{rjags::coda.samples}}
}
\value{
array of summary statistics for \code{p} and \code{k}
}
\description{
Summary statistics for estimated p and p parameters
}
